User activity details
- Displaying information about user events on user info screen
- Refactored user info tabs (added abstraction)
- TODO: Starred tab, Followers/Following infos

pszklarska/FlutterGitHub#1
