+++ lib/data/model/event.dart
@@ -4,8 +4,9 @@ class Event {
   final EventType type;
   final User actor;
   final EventPayload payload;
+  final EventRepo repo;
 
-  Event(this.type, this.actor, this.payload);
+  Event(this.type, this.actor, this.payload, this.repo);
 
   factory Event.fromJson(json) {
     if (json == null) {
@@ -14,7 +15,9 @@ class Event {
       return new Event(
           getEventType(json['type']),
           new User.fromJson(json['actor']),
-          new EventPayload.fromJson(json['payload']));
+          new EventPayload.fromJson(json['payload']),
+          new EventRepo.fromJson(json['repo'])
+      );
     }
   }
 }
@@ -55,6 +58,21 @@ class EventPayload {
       return null;
     } else {
       return new EventPayload(json['action'], json['size'], json['ref_type']);
+    }
+  }
+}
+
+class EventRepo {
+  final String url;
+  final String name;
+
+  EventRepo(this.url, this.name);
+
+  factory EventRepo.fromJson(json){
+    if (json == null) {
+      return null;
+    } else {
+      return new EventRepo(json['url'], json['name']);
     }
   }
 }
