+++ lib/ui/user_info/tabs/profile_tab.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/cupertino.dart';
+
+abstract class ProfileTab extends StatelessWidget {
+  final String title;
+  final IconData icon;
+
+  ProfileTab(this.title, this.icon);
+}
